cmd/link: generate R_X86_64_RELATIVE for internal PIE R_ADDR
Currently internal PIE generates a lot of relocs.

For instance, PIE cmd/go, built at HEAD with:

    $ go build -o /tmp/go -buildmode=pie -ldflags=-linkmode=internal -a cmd/go

This generates a PIE go tool with 75557 relocations, the vast majority
of which are R_X86_64_64 [1]. These are not very efficient for the
dynamic linker to handle, so a simple 'go version' with this binary
takes 2.9 seconds (!) to start on my machine.

Switching to the more efficient R_X86_64_RELATIVE is fairly trivial. We
just need an additional Addaddrplus to resolve targ in reloc() and we're
good to go.

With this change, PIE 'go version' takes ~50ms on my machine (basically
indistinguishable from the statically linked version). It also reduces
the binary size from 17MB to 14MB by avoiding Adddynsym.

R_X86_64_64 are still generated for SDYNIMPORT symbols, since those are
truly external.

[1] It's 100% X86_64_64 without cgo. Ideally the compiler would be
generating mostly PC relative relocs that the linker could resolve
internally, but that is a much more involved change.

Change-Id: I955764fe6b545a47ab2feb3979a345d43df682c2